feat(Coin): 코인의 amount 리스트를 반환하는 기능 구현
